Support `{{ item|markup }}` for outputting escaped content in loops
It's currently possible to use the syntax `{{ item|_markup }}` to output an
array's content with proper escaping, but a pattern that does so cannot be
viewed in the Larva server as Twing reports
`rawMessage: 'Unknown "_markup" filter.'`. This makes `{{ item|markup }}` the
correct way to escape an array element within the loop syntax we support
without requiring the use of an action.