Support webpack externals in Larva config
If a Larva script will use a package provided by WordPress, such as lodash or
any `@wordpress`-prefixed package, it should be declared as an external in the
Larva config otherwise webpack will either error because it can't resolve the
module or will require that we bundle one of those packages in the module. By
allowing consuming projects to specify externals, those packages can be
sourced from WordPress.